# improves sigverify discard_excess_packets performance