#### Problem
As shown by the added benchmark, current code does worse if there is a
spam address plus a lot of unique addresses.

on current master:
```
test bench_packet_discard_many_senders  ... bench:   1,997,960 ns/iter (+/- 103,715)
test bench_packet_discard_mixed_senders ... bench:  14,256,116 ns/iter (+/- 534,865)
test bench_packet_discard_single_sender ... bench:   1,306,809 ns/iter (+/- 61,992)
```

with this commit:
```
test bench_packet_discard_many_senders  ... bench:   1,644,025 ns/iter (+/- 83,715)
test bench_packet_discard_mixed_senders ... bench:   1,089,789 ns/iter (+/- 86,324)
test bench_packet_discard_single_sender ... bench:     955,234 ns/iter (+/- 55,953)
```
#### Summary of Changes
* No change to the order of packets or which packets are discarded.
* Drop hash-map entries if there are no more packets from respective address.
* Allocate `max_packets` with fewer iterations.